refactor(system): extract ControllerPorts to separate module
Extract ControllerPorts struct and implementation from system.rs
to improve modularity and code organization.

Changes:
- Create src/core/system/ directory structure
- Move system.rs to src/core/system/mod.rs
- Extract ControllerPorts to controller_ports.rs (238 lines)
- Remove unused Controller import from system module
- Add module declaration and re-export

Benefits:
- Reduced system module size from 1,782 to 1,567 lines (~215 lines)
- Better separation of concerns
- Improved code discoverability
- Easier to test ControllerPorts in isolation
- Consistent with existing architecture (cpu/, gpu/, memory/)

All 543 tests pass. No functional changes.

Closes #148
